+++ packages/dash-table/CHANGELOG.md
@@ -11,6 +11,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
     - Clicking on a link in a Markdown cell now requires a single click instead of two
     - Links in Markdown cells now open a new tab (target="_blank")
 
+### Fixed
+- [#785](https://github.com/plotly/dash-table/pull/785) Fix a bug where the table does not refresh correctly if a property was previously missing
+
 ## [4.7.0] - 2020-05-05
 ### Added
 - [#729](https://github.com/plotly/dash-table/pull/729) Improve conditional styling