+++ packages/dash-table/CHANGELOG.md
@@ -13,6 +13,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - [#785](https://github.com/plotly/dash-table/pull/785) Fix a bug where the table does not refresh correctly if a property was previously missing
+- [#793](https://github.com/plotly/dash-table/pull/793)
+    - Fix a bug where headers aren't aligned with columns with fixed_rows [#777](https://github.com/plotly/dash-table/issues/777)
+    - Fix a regression where headers don't scroll horizontally with fixed_rows [#780](https://github.com/plotly/dash-table/issues/780)
 
 ## [4.7.0] - 2020-05-05
 ### Added